requirements.txt
# Day Trade Content Generator - Requirements
# Core dependencies
requests==2.31.0
schedule==1.2.2
python-dotenv==1.0.0

# Google APIs
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0

# Data processing
pandas==2.1.1
openpyxl==3.1.2

# Web framework (para webhooks opcionais)
flask==2.3.3
gunicorn==21.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
urllib3==2.0.4
